Add files via upload
The memory formation is linked to changes the strength of synaptic connections between neurons in the brain. The classical hebbian learning approach is unstable, since it would drive the synaptic strengths to saturation. The neural network will then loose its efficacy to encode new information. Although the brain has about 100 billion neurons, each of which connected to about 10,000 others (depending on part of brain this number may vary), we still have limited storage as compared to the information the brain processes everyday. The brain has an elegant mechanism by which it can alter the synaptic strengths whenever required. The role of Ca2+ ion mediated by the NMDA receptor is well studied. This model takes the physical dynamics of the NMDA receptor and proposes a way in which brain implements metalasticity