Role of RyRs in metaplasticity in dendritic spines
The code uses biologically realistic models of various ion channels associated in a dendritic spine to model its behavior at large. The major part of code (except for RyR modelling) is written by Gaurang Mahajan. Through this study, it is intended to find the role of RyR as a ER store in mediating spine specific plasticity.